Sligthly changed the API to remove unnecessary unsigned int -> double transformation.

+++ src/CH/CHDistanceQueryManagerWithMapping.cpp
@@ -15,6 +15,6 @@ CHDistanceQueryManagerWithMapping::CHDistanceQueryManagerWithMapping(FlagsGraph 
 
 // Used to find the shortest distance from source to target where source and target are the original indices.
 //______________________________________________________________________________________________________________________
-double CHDistanceQueryManagerWithMapping::findDistance(const long long unsigned int source, const long long unsigned int target) {
+unsigned int CHDistanceQueryManagerWithMapping::findDistance(const long long unsigned int source, const long long unsigned int target) {
     return qm.findDistance(mapping.at(source), mapping.at(target));
 }
